bot/requirements.txt
python-telegram-bot==21.5
aiohttp==3.9.5
aiosqlite==0.20.0
python-dotenv==1.0.1
apscheduler==3.10.4
